Performance: clear key material with word-wide stores; fix UB in sub_word
The per-call key-material wipe (expanded key plus loaded round keys, ~500 B
per call) was a byte-wise volatile loop costing ~1.7 ns/byte. On small
payloads -- one AES block per MTProto message -- that dominated the call: a
16-byte IGE encrypt spent well over half of its time in the wipe.

Clear whole 4-byte words where the pointer allows it instead (align up with at
most 3 byte stores, then word stores, then a byte tail). Same guarantee, the
stores are still volatile so nothing can be optimized away, for ~8x fewer
stores. Measured on a Celeron N4120 (AES-NI path, best of 5, C call only, so
no Python overhead included):

  ige enc 16 B    1742 ->  984 ns
  ige dec 16 B    1407 ->  559 ns
  ctr enc 16 B    1401 ->  713 ns
  ige enc 1 KiB   4692 -> 3510 ns

Also fix undefined behaviour in sub_word(): the SBOX bytes were left-shifted
as promoted int (e.g. 0xA4 << 24), overflowing the sign bit; UBSAN flagged it.
Shifting unsigned values produces exactly the same key schedule.

Validation:
- byte-for-byte identical output vs. the previous build: the 63-case
  differential and the 255-case AES-NI/software cross-path differential are
  unchanged
- 3886 tests pass with AES-NI and with TGCRYPTO_NO_AESNI
- ASAN+UBSAN clean, including a wipe alignment/size matrix (sizes 0-64,
  offsets 0-7) asserting exactly [p, p+n) is zeroed and neighbours untouched

+++ tgcrypto/aes256.c
@@ -350,11 +350,13 @@ static const uint32_t RCON[10] = {
 };
 
 uint32_t sub_word(uint32_t word) {
+    /* The shifts are performed on unsigned values: shifting a promoted int
+     * that overflows the sign bit (e.g. 0xA4 << 24) is undefined behaviour. */
     return (
-        (SBOX[word >> 4 & 0x0f][word & 0xf])
-        + (SBOX[word >> 12 & 0x0f][word >> 8 & 0x0f] << 8)
-        + (SBOX[word >> 20 & 0x0f][word >> 16 & 0x0f] << 16)
-        + (SBOX[word >> 28 & 0x0f][word >> 24 & 0x0f] << 24)
+        (uint32_t) SBOX[word >> 4 & 0x0f][word & 0xf]
+        + ((uint32_t) SBOX[word >> 12 & 0x0f][word >> 8 & 0x0f] << 8)
+        + ((uint32_t) SBOX[word >> 20 & 0x0f][word >> 16 & 0x0f] << 16)
+        + ((uint32_t) SBOX[word >> 28 & 0x0f][word >> 24 & 0x0f] << 24)
     );
 }
